python/requirements.txt
pyserial-asyncio>=0.6
pandas>=1.5.0
crcmod>=1.7
pyserial>=3.5
PyYAML>=6.0
matplotlib>=3.6.0
tkinter>=8.6

